requirements.txt
asgiref==3.5.2
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.5
Django==2.2.16
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.11.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest-pythonpath==0.7.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
zipp==3.8.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0
